requirements.txt
aerich==0.8.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiomysql==0.2.0
aiosignal==1.3.2
aiosmtplib==3.0.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dictdiffer==0.9.0
dnspython==2.7.0
email_validator==2.2.0
Faker==37.1.0
fastapi==0.115.12
fastapi-mail==1.4.2
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
multidict==6.3.2
narwhals==1.35.0
nextcord==2.6.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==5.29.4
psycopg2==2.9.10
pyarrow==19.0.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
PyMySQL==1.1.1
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
streamlit==1.44.1
stripe==12.0.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tortoise-orm==0.24.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
